portfolio/serializers.py
from rest_framework import serializers
from .models import Project, Skill, ContactMessage

class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = '__all__'

class SkillSerializer(serializers.ModelSerializer):
    class Meta:
        model = Skill
        fields = '__all__'

class ContactMessageSerializer(serializers.ModelSerializer):
    from .models import ContactMessage

    class Meta:
        model = ContactMessage
        fields = '__all__'        